remove VideoPlayer and switch to VideoController for cache
